+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/zombiezen/gg-git/compare/v0.2.0...HEAD
 
+### Added
+
+-  Add `ParseURL` function
+
 ## [0.2.0][] - 2020-08-17
 
 Version 0.2.0 adds functionality for inspecting [submodules][].